# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/Repository/DirectorRepositoryImpl.java
#	src/main/java/ru/yandex/practicum/filmorate/Repository/RatingRepositoryImpl.java
#	src/main/java/ru/yandex/practicum/filmorate/Service/DirectorServiceImpl.java
#	src/main/java/ru/yandex/practicum/filmorate/Service/FilmServiceImpl.java
#	src/main/java/ru/yandex/practicum/filmorate/Service/MpaServiceImpl.java
#	src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/FilmGenreRowMapper.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/FilmRowMapper.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/FriendshipRowMapper.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/GenreRowMapper.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/LikesRowMapper.java
#	src/main/java/ru/yandex/practicum/filmorate/dal/mappers/UserRowMapper.java

+++ src/main/java/ru/yandex/practicum/filmorate/Service/DirectorServiceImpl.java
@@ -5,11 +5,11 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.Repository.DirectorRepository;
 import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Director;
-
 import java.util.Collection;
 
 @Service
 @RequiredArgsConstructor
+public class DirectorServiceImpl implements DirectorService {
 public class DirectorImpl implements DirectorService {
     private final DirectorRepository directorRepository;
 
@@ -31,11 +31,17 @@ public class DirectorImpl implements DirectorService {
 
     @Override
     public Director update(Director director) {
+        if (directorRepository.findById(director.getId()).isEmpty()) {
+            throw new NotFoundException("No director with id: " + director.getId());
+        }
         return directorRepository.update(director);
     }
 
     @Override
     public boolean delete(int directorId) {
+        if (directorRepository.findById(directorId).isEmpty()) {
+            throw new NotFoundException("No director with id: " + directorId);
+        }
         return directorRepository.delete(directorId);
     }
 }
